requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
httpx>=0.25.0
google-auth>=2.22.0
google-generativeai>=0.3.1
python-dotenv>=1.0.0
email-validator>=2.0.0
PyPDF2>=3.0.0
mammoth>=1.6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.28.0
pydantic-settings>=2.0.3 